Move WiremockAwsV2ClientHelper to localstack-test module
